Fix for iPad for error:
Your application has presented a UIAlertController (<UIAlertController: 0x7fa4e56dda00>) of style
UIAlertControllerStyleActionSheet from UIViewController (<UIViewController: 0x7fa500c59390>). The modalPresentationStyle of a UIAlertController with this style is UIModalPresentationPopover. You must provide
location information for this popover through the alert controller's popoverPresentationController. You must provide either a sourceView and sourceRect or a barButtonItem.  If this information is not known when
you present the alert controller, you may provide it in the UIPopoverPresentationControllerDelegate method -prepareForPopoverPresentation.